Specify sketches report input values for sketch compilation CI actions
The reliance on the `actions/upload-artifacts` action as the transfer mechanism between workflow runs means that these
values will always be hardcoded into the workflow. The result is that if the default values of the inputs are changed, it
breaks the workflow, even if the changes are synchronized. So it's more robust to avoid dependening on the default
values.